+++ CHANGELOG.md
@@ -10,7 +10,9 @@
   - Grasshopper Chess (`FairyVariants.grasshopper()`): chess with some extra grasshopper pieces.
   - Berolina Chess (`FairyVariants.berolina()`): chess with the Berolina pawn (`PieceType.berolinaPawn()`).
 - Betza parsing now supports `(x,y)` atoms, allowing for pieces jumping more than 3 squares in each direction. For example, the 'Giraffe' is `(4,1)`, and the Knight could be rewritten as `(2,1)` (or `(1,2)`).
+- `FirstMoveOptions`: replaces `Variant.firstMoveRanks`, supports more flexible first move options.
 - A lot of old deprecated constants and functions have been removed.
+- The default value of `Variant.castlingOptions` is now `CastlingOptions.none`.
 - `play.dart` example: move filtering commands - `moves from x`, `moves to x`, `moves captures`, `moves quiet`.
 - `BoardSize.isValidSquareName()`: check if a square name is valid.